Remove unused static files from the project
A bulk of static files were removed in this commit. These files primarily include aspects of the admin interface design and JavaScript code, as well as the i18n localization data. They were no longer needed or used in the application, hence the removal, which leads to cleaner project structure and lower overall project size.