Refactor model declarations with improved structure
The model classes and admin site registrations have been shuffled between accounts and dash applications to follow a more logical structure. This commit moves Account, UserSetting, CreditCard, and Transaction models from the dash app to the accounts app and updates the corresponding imports in the admin. The admin registrations for these models are accordingly updated in both apps. Furthermore, Password and Port fields for the database connection in settings.py were modified.