Restructure home page and update user login form handling
This commit includes major change in the structure and content of the home.html template. Specifically, the login form is now only displayed if the user is not authenticated. Additionally, placeholder texts and headings have been replaced with relevant banking-related content to provide a more complete and user-friendly interface. Unnecessary changes in workspace.xml have been cleaned up as well.